+++ src/main/java/co/edu/unbosque/veterinaria/repository/AdoptanteRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdoptanteRepository extends CrudRepository<Adoptante, Long> {
-}
+public interface AdoptanteRepository extends CrudRepository<Adoptante, Integer> { }
