+++ src/main/java/co/edu/unbosque/veterinaria/repository/AuditoriaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuditoriaRepository extends CrudRepository<Auditoria, Long> {
+public interface AuditoriaRepository extends CrudRepository<Auditoria, Integer> {
 }
