+++ src/main/java/co/edu/unbosque/veterinaria/repository/CuidadorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CuidadorRepository extends CrudRepository<Cuidador, Long> {
+public interface CuidadorRepository extends CrudRepository<Cuidador, Integer> {
 }
