+++ src/main/java/co/edu/unbosque/veterinaria/repository/MascotaRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MascotaRepository extends CrudRepository<Mascota, Long> {
-}
+public interface MascotaRepository extends CrudRepository<Mascota, Integer> { }
