+++ src/main/java/co/edu/unbosque/veterinaria/repository/UsuarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import co.edu.unbosque.veterinaria.entity.Usuario;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
 }
