+++ src/main/java/co/edu/unbosque/veterinaria/repository/VacunaCatalogoRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VacunaCatalogoRepository extends CrudRepository<VacunaCatalogo, Long> {
-}
+public interface VacunaCatalogoRepository extends CrudRepository<VacunaCatalogo, Integer> { }
