+++ src/main/java/co/edu/unbosque/veterinaria/service/api/AdoptanteServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Adoptante;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface AdoptanteServiceAPI extends GenericServiceAPI<Adoptante, Long> { }
+public interface AdoptanteServiceAPI extends GenericServiceAPI<Adoptante, Integer> { }
