+++ src/main/java/co/edu/unbosque/veterinaria/service/api/EspecieServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Especie;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface EspecieServiceAPI extends GenericServiceAPI<Especie, Long> { }
+public interface EspecieServiceAPI extends GenericServiceAPI<Especie, Integer> { }
