+++ src/main/java/co/edu/unbosque/veterinaria/service/api/HistorialMedicoServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.HistorialMedico;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface HistorialMedicoServiceAPI extends GenericServiceAPI<HistorialMedico, Long> { }
+public interface HistorialMedicoServiceAPI extends GenericServiceAPI<HistorialMedico, Integer> { }
