+++ src/main/java/co/edu/unbosque/veterinaria/service/api/MascotaServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Mascota;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface MascotaServiceAPI extends GenericServiceAPI<Mascota, Long> { }
+public interface MascotaServiceAPI extends GenericServiceAPI<Mascota, Integer> { }
