+++ src/main/java/co/edu/unbosque/veterinaria/service/api/UsuarioServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Usuario;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface UsuarioServiceAPI extends GenericServiceAPI<Usuario, Long> { }
+public interface UsuarioServiceAPI extends GenericServiceAPI<Usuario, Integer> { }
