+++ src/main/java/co/edu/unbosque/veterinaria/service/api/VacunaCatalogoServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.VacunaCatalogo;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface VacunaCatalogoServiceAPI extends GenericServiceAPI<VacunaCatalogo, Long> { }
+public interface VacunaCatalogoServiceAPI extends GenericServiceAPI<VacunaCatalogo, Integer> { }
