+++ src/main/java/co/edu/unbosque/veterinaria/service/impl/MascotaServiceImpl.java
@@ -9,12 +9,13 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class MascotaServiceImpl extends GenericServiceImpl<Mascota, Long> implements MascotaServiceAPI {
+public class MascotaServiceImpl extends GenericServiceImpl<Mascota, Integer> implements MascotaServiceAPI {
+
     @Autowired
     private MascotaRepository mascotaRepository;
 
     @Override
-    public CrudRepository<Mascota, Long> getDao() {
+    public CrudRepository<Mascota, Integer> getDao() {
         return mascotaRepository;
     }
 }
